LayoutTests/crypto/subtle/verify-malformed-parameters-expected.txt

Test verification with malformed parameters

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS crypto.subtle.verify() rejected promise  with TypeError: Not enough arguments.
PASS crypto.subtle.verify(1) rejected promise  with TypeError: Not enough arguments.
PASS crypto.subtle.verify(1, 2) rejected promise  with TypeError: Not enough arguments.
PASS crypto.subtle.verify(1, 2, 3) rejected promise  with TypeError: Not enough arguments.
PASS crypto.subtle.verify("RSA-PSS", wrongKey, signature, text) rejected promise  with TypeError: Member RsaPssParams.saltLength is required and must be an instance of unsigned long.
PASS crypto.subtle.verify("hmac", wrongKey, signature, text) rejected promise  with InvalidAccessError: CryptoKey doesn't support verification.
PASS successfullyParsed is true

TEST COMPLETE
